did a bit of rearrangement, reduced public interface a bit
