+++ src/ctrlc_breaker.rs
@@ -3,7 +3,8 @@ use futures::prelude::*;
 use futures::sync;
 use organelle::{Axon, Impulse, Soma};
 
-use super::{Error, Result, Synapse};
+use super::{Error, Result};
+use synapse::Synapse;
 
 /// soma that stops the organelle upon Ctrl-C
 pub struct CtrlcBreakerSoma;
